Limiting file handle expressions between 1 and 255 during parsing
Runtime error if an integer expression that is supposed to be a file handle exceeds the 1 - 255 bounds

Support CLOSE integer-expression

Support parenthesis expressions in OPEN

+++ src/common/file_constants.rs
@@ -59,7 +59,7 @@ impl From<i32> for FileAccess {
 }
 
 #[derive(Clone, Copy, Debug, Default, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct FileHandle(u32);
+pub struct FileHandle(u8);
 
 impl FileHandle {
     pub fn is_valid(&self) -> bool {
@@ -67,8 +67,8 @@ impl FileHandle {
     }
 }
 
-impl From<u32> for FileHandle {
-    fn from(x: u32) -> FileHandle {
+impl From<u8> for FileHandle {
+    fn from(x: u8) -> FileHandle {
         FileHandle(x)
     }
 }
